back-end/pyproject.toml
[project]
name = "sunhacks-25"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "dotenv>=0.9.9",
    "fastapi>=0.117.1",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.0.10",
    "langchain-tavily>=0.2.11",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.11.0",
    "pydantic[email]>=2.11.9",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy[asyncio]>=2.0.43",
    "tavily-python>=0.7.12",
    "uvicorn[standard]>=0.37.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]
